+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 ## [Unreleased]
 None
 
+## [1.1.0] - 2023-09-13
+### Added
+- Multi line support
+- Create without switching to source file
+### Fixed
+- more reliable classname
+- Removed errant spaces
+
 ## [1.0.6] - 2023-05-06
 ### Added
 - Docs Example pic